+++ test/unittests/test_create_project.py
@@ -4,8 +4,8 @@ import unittest
 from typing import Any
 
 from dsp_tools.models.helpers import BaseError
-from dsp_tools.utils.onto_create_ontology import _sort_resources, _sort_prop_classes
-from dsp_tools.utils.onto_validate import _collect_link_properties, _identify_problematic_cardinalities, \
+from dsp_tools.utils.project_create import _sort_resources, _sort_prop_classes
+from dsp_tools.utils.project_validate import _collect_link_properties, _identify_problematic_cardinalities, \
     validate_project
